Enable import of the training set by default.
With the new signature of ReductionMethod.offline(), which returns a reduced order parametrized problem, we automatically enable import of previous offline stages, if they were already computed. Thus, it does not hurt to be consistent also in the import of the training set, which now defaults to True.
To start again from scratch, the user will need to consciously remove all folders generated by RBniCS (traning set folder included).
The default value for the import of the test set reamins to False, because in multiple runs the user will want to test the reduced order method for randomly generated test sets, which will change at every run.