Define new extends(Parent, preserve_class_name=False) and override decorators.
These decorators, not yet implemented, should help in getting the doc string from parent class, if the child class does not provide one. Thanks to this, help() will work properly. Having @override for each method which was overridden in the child may also help the user when reading the code. Note, however, that when running the code these decorators are (almost) completely optional, meaning that they do not add any real code: indeed we do not use them in the tutorials, even though they are applicable there too. The exception to the previous statement is the case when attribute preserve_class_name is True: this is the same as the former KeepClassName decorator, which has been removed. This is enabled for EIM, SCM and ShapeParametrization in order to have more meaningful folder names and nicer python errors (e.g. AttributeError in MyEllipticProblem rather than in EIMDecoratedParametrizedProblem_Class).