Several methods, in particular in linear algebra, have the following structure assert(all implemented cases) if (case 1): 	do case 1 elif (case 2): 	do case 2 elif ... 	... else: 	return Error("...")
Change the type of the raised Error to be an AssertionError since it is basically a duplicate of the assert on top. It is helpful in any case to leave the explicit else at the end, to avoid mistakes in future (e.g. adding the case to the assert but not to the if)